refactor: :recycle: simplify implementation
